+++ src/constants.ts
@@ -1,0 +1,8 @@
+/**
+ * Extension constants and defaults
+ */
+
+export const DEBUG = false;  // general debug
+
+// FIXME: make configurable?
+export const MAX_TIMEOUT = 1000;  // msec
